requirements.txt
APScheduler==3.6.3
click==7.1.2
email-validator==1.1.2
fastapi==0.63.0
fastapi-mail==0.3.4.2 
Jinja2==3.0.3 
passlib==1.7.4
psycopg2==2.8.6
pydantic==1.7.3
pyparsing==2.4.6
python-multipart==0.0.5
pytz==2021.1 
sniffio==1.2.0
SQLAlchemy==1.3.23 
starlette==0.13.6 
tzlocal==2.1
urllib3==1.25.8
uvicorn==0.13.3 
fastapi-login==1.5.2
python-jose==3.2.0 
jose==1.0.0
bcrypt==3.2.0 
aiosqlite==0.16.0
h11==0.10.0
python-dateutil==2.8.1
six==1.15.0
python-dotenv==0.15.0
gunicorn==19.9.0
httptools==0.1.1
uvloop==0.14.0
databases==0.4.1
psycopg2-binary==2.8.6
asyncpg==0.21.0
PyJWT==1.4.0
aiofiles==0.6.0
email-validator==1.1.2 
